test(scoring): commodity ceiling + freshness multiplier regression tests
7 new pipeline tests locking in calibration behavior:
- Tutorial/HelpRequest capped at 0.28/0.22 when sophistication < 0.35
- High sophistication (>= 0.35) bypasses commodity ceiling
- CVE title pattern bypasses commodity ceiling
- Fresh items (1h) get freshness_mult > 1.0
- 72h items are neutral (freshness_mult ~= 1.0)
- Old items (720h) get freshness_mult < 0.85

Also fixes 2 Clippy warnings: unnecessary f32 cast in pipeline_v2,
dead code in intelligence_metrics (moved to #[cfg(test)]).

+++ src-tauri/src/scoring/pipeline_v2.rs
@@ -875,8 +875,7 @@ fn compute_quality_composite(
     // Negative stack prior: Bayesian suppression for technologies user doesn't use.
     // UNDAMPENED — full suppressive force (0.15 for competing-absent, 0.30 for anti-topics).
     let negative_stack_prior =
-        crate::stacks::negative_stack::lookup_prior(&ctx.ace_ctx.negative_stack, &raw.topics)
-            as f32;
+        crate::stacks::negative_stack::lookup_prior(&ctx.ace_ctx.negative_stack, &raw.topics);
 
     // NOTE: ecosystem_shift_mult, stack_competing_mult, and content_analysis_mult are
     // still computed above for the return tuple (used by logging/diagnostics) but are
